Split CLI files
