+++ cmd/system-probe/subcommands/runtime/discarders.go
@@ -1,0 +1,65 @@
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the Apache License Version 2.0.
+// This product includes software developed at Datadog (https://www.datadoghq.com/).
+// Copyright 2016-present Datadog, Inc.
+
+//go:build linux || windows
+
+// Package runtime holds runtime related files
+package runtime
+
+import (
+	"fmt"
+
+	"github.com/DataDog/datadog-agent/cmd/system-probe/command"
+	"github.com/DataDog/datadog-agent/comp/core"
+	"github.com/DataDog/datadog-agent/comp/core/config"
+	log "github.com/DataDog/datadog-agent/comp/core/log/def"
+	"github.com/DataDog/datadog-agent/comp/core/secrets"
+	secagent "github.com/DataDog/datadog-agent/pkg/security/agent"
+	"github.com/DataDog/datadog-agent/pkg/util/fxutil"
+	"github.com/spf13/cobra"
+	"go.uber.org/fx"
+)
+
+func discardersCommands(_ *command.GlobalParams) []*cobra.Command {
+	dumpDiscardersCmd := &cobra.Command{
+		Use:   "dump",
+		Short: "dump discarders",
+		RunE: func(_ *cobra.Command, _ []string) error {
+			return fxutil.OneShot(dumpDiscarders,
+				fx.Supply(core.BundleParams{
+					ConfigParams: config.NewAgentParams("", config.WithConfigMissingOK(true)),
+					SecretParams: secrets.NewDisabledParams(),
+					LogParams:    log.ForOneShot("SYS-PROBE", "info", true)}),
+				core.Bundle(),
+			)
+		},
+	}
+
+	discardersCmd := &cobra.Command{
+		Use:   "discarders",
+		Short: "discarders commands",
+	}
+	discardersCmd.AddCommand(dumpDiscardersCmd)
+
+	return []*cobra.Command{discardersCmd}
+}
+
+func dumpDiscarders(_ log.Component, _ config.Component, _ secrets.Component) error {
+	runtimeSecurityClient, err := secagent.NewRuntimeSecurityClient()
+	if err != nil {
+		return fmt.Errorf("unable to create a runtime security client instance: %w", err)
+	}
+	defer runtimeSecurityClient.Close()
+
+	dumpFilename, dumpErr := runtimeSecurityClient.DumpDiscarders()
+
+	if dumpErr != nil {
+		return fmt.Errorf("unable to dump discarders: %w", dumpErr)
+	}
+
+	fmt.Printf("Discarder dump file: %s\n", dumpFilename)
+
+	return nil
+}
