+++ cmd/system-probe/subcommands/runtime/discarders_test.go
@@ -1,0 +1,24 @@
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the Apache License Version 2.0.
+// This product includes software developed at Datadog (https://www.datadoghq.com/).
+// Copyright 2016-present Datadog, Inc.
+
+//go:build linux || windows
+
+// Package runtime holds runtime related files
+package runtime
+
+import (
+	"testing"
+
+	"github.com/DataDog/datadog-agent/cmd/system-probe/command"
+	"github.com/DataDog/datadog-agent/pkg/util/fxutil"
+)
+
+func TestDumpDiscardersCommand(t *testing.T) {
+	fxutil.TestOneShotSubcommand(t,
+		Commands(&command.GlobalParams{}),
+		[]string{"runtime", "discarders", "dump"},
+		dumpDiscarders,
+		func() {})
+}
