+++ cmd/system-probe/subcommands/runtime/network_namespace.go
@@ -1,0 +1,80 @@
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the Apache License Version 2.0.
+// This product includes software developed at Datadog (https://www.datadoghq.com/).
+// Copyright 2016-present Datadog, Inc.
+
+//go:build linux
+
+// Package runtime holds runtime related files
+package runtime
+
+import (
+	"fmt"
+
+	"github.com/DataDog/datadog-agent/cmd/system-probe/command"
+	"github.com/DataDog/datadog-agent/comp/core"
+	"github.com/DataDog/datadog-agent/comp/core/config"
+	log "github.com/DataDog/datadog-agent/comp/core/log/def"
+	"github.com/DataDog/datadog-agent/comp/core/secrets"
+	secagent "github.com/DataDog/datadog-agent/pkg/security/agent"
+	"github.com/DataDog/datadog-agent/pkg/util/fxutil"
+	"github.com/spf13/cobra"
+	"go.uber.org/fx"
+)
+
+type dumpNetworkNamespaceCliParams struct {
+	*command.GlobalParams
+
+	snapshotInterfaces bool
+}
+
+func networkNamespaceCommands(globalParams *command.GlobalParams) []*cobra.Command {
+	cliParams := &dumpNetworkNamespaceCliParams{
+		GlobalParams: globalParams,
+	}
+
+	dumpNetworkNamespaceCmd := &cobra.Command{
+		Use:   "dump",
+		Short: "dumps the network namespaces held in cache",
+		RunE: func(_ *cobra.Command, _ []string) error {
+			return fxutil.OneShot(dumpNetworkNamespace,
+				fx.Supply(cliParams),
+				fx.Supply(core.BundleParams{
+					ConfigParams: config.NewAgentParams("", config.WithConfigMissingOK(true)),
+					SecretParams: secrets.NewDisabledParams(),
+					LogParams:    log.ForOneShot("SYS-PROBE", "info", true)}),
+				core.Bundle(),
+			)
+		},
+	}
+	dumpNetworkNamespaceCmd.Flags().BoolVar(&cliParams.snapshotInterfaces, "snapshot-interfaces", true, "snapshot the interfaces of each network namespace during the dump")
+
+	networkNamespaceCmd := &cobra.Command{
+		Use:   "network-namespace",
+		Short: "network namespace command",
+	}
+	networkNamespaceCmd.AddCommand(dumpNetworkNamespaceCmd)
+
+	return []*cobra.Command{networkNamespaceCmd}
+}
+
+func dumpNetworkNamespace(_ log.Component, _ config.Component, _ secrets.Component, dumpNetworkNamespaceArgs *dumpNetworkNamespaceCliParams) error {
+	client, err := secagent.NewRuntimeSecurityClient()
+	if err != nil {
+		return fmt.Errorf("unable to create a runtime security client instance: %w", err)
+	}
+	defer client.Close()
+
+	resp, err := client.DumpNetworkNamespace(dumpNetworkNamespaceArgs.snapshotInterfaces)
+	if err != nil {
+		return fmt.Errorf("couldn't send network namespace cache dump request: %w", err)
+	}
+
+	if len(resp.GetError()) > 0 {
+		return fmt.Errorf("couldn't dump network namespaces: %s", resp.GetError())
+	}
+
+	fmt.Printf("Network namespace dump: %s\n", resp.GetDumpFilename())
+	fmt.Printf("Network namespace dump graph: %s\n", resp.GetGraphFilename())
+	return nil
+}
