+++ cmd/system-probe/subcommands/runtime/network_namespace_test.go
@@ -1,0 +1,24 @@
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the Apache License Version 2.0.
+// This product includes software developed at Datadog (https://www.datadoghq.com/).
+// Copyright 2016-present Datadog, Inc.
+
+//go:build linux
+
+// Package runtime holds runtime related files
+package runtime
+
+import (
+	"testing"
+
+	"github.com/DataDog/datadog-agent/cmd/system-probe/command"
+	"github.com/DataDog/datadog-agent/pkg/util/fxutil"
+)
+
+func TestDumpNetworkNamespaceCommand(t *testing.T) {
+	fxutil.TestOneShotSubcommand(t,
+		Commands(&command.GlobalParams{}),
+		[]string{"runtime", "network-namespace", "dump"},
+		dumpNetworkNamespace,
+		func() {})
+}
