+++ cmd/system-probe/subcommands/runtime/policy/eval_test.go
@@ -1,0 +1,26 @@
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the Apache License Version 2.0.
+// This product includes software developed at Datadog (https://www.datadoghq.com/).
+// Copyright 2016-present Datadog, Inc.
+
+//go:build linux || windows
+
+// Package policy holds policy CLI subcommand related files
+package policy
+
+import (
+	"testing"
+
+	"github.com/DataDog/datadog-agent/cmd/system-probe/command"
+	"github.com/DataDog/datadog-agent/pkg/util/fxutil"
+
+	"github.com/spf13/cobra"
+)
+
+func TestEvalCommand(t *testing.T) {
+	fxutil.TestOneShotSubcommand(t,
+		[]*cobra.Command{EvalCommand(&command.GlobalParams{})},
+		[]string{"runtime", "policy", "eval", "--rule-id=10", "--event-file=file"},
+		evalRule,
+		func() {})
+}
