+++ cmd/system-probe/subcommands/runtime/policy/reload.go
@@ -1,0 +1,56 @@
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the Apache License Version 2.0.
+// This product includes software developed at Datadog (https://www.datadoghq.com/).
+// Copyright 2016-present Datadog, Inc.
+
+//go:build linux || windows
+
+// Package policy holds policy CLI subcommand related files
+package policy
+
+import (
+	"fmt"
+
+	"github.com/DataDog/datadog-agent/cmd/system-probe/command"
+	"github.com/DataDog/datadog-agent/comp/core"
+	"github.com/DataDog/datadog-agent/comp/core/config"
+	log "github.com/DataDog/datadog-agent/comp/core/log/def"
+	"github.com/DataDog/datadog-agent/comp/core/secrets"
+	secagent "github.com/DataDog/datadog-agent/pkg/security/agent"
+	"github.com/DataDog/datadog-agent/pkg/util/fxutil"
+
+	"github.com/spf13/cobra"
+	"go.uber.org/fx"
+)
+
+// ReloadPoliciesCommand returns the CLI command for "policy reload"
+func ReloadPoliciesCommand(_ *command.GlobalParams) *cobra.Command {
+	return &cobra.Command{
+		Use:   "reload",
+		Short: "Reload policies",
+		RunE: func(_ *cobra.Command, _ []string) error {
+			return fxutil.OneShot(reloadRuntimePolicies,
+				fx.Supply(core.BundleParams{
+					ConfigParams: config.NewAgentParams("", config.WithConfigMissingOK(true)),
+					SecretParams: secrets.NewDisabledParams(),
+					LogParams:    log.ForOneShot("SYS-PROBE", "info", true)}),
+				core.Bundle(),
+			)
+		},
+	}
+}
+
+func reloadRuntimePolicies(_ log.Component, _ config.Component, _ secrets.Component) error {
+	client, err := secagent.NewRuntimeSecurityClient()
+	if err != nil {
+		return fmt.Errorf("unable to create a runtime security client instance: %w", err)
+	}
+	defer client.Close()
+
+	_, err = client.ReloadPolicies()
+	if err != nil {
+		return fmt.Errorf("unable to reload policies: %w", err)
+	}
+
+	return nil
+}
