+++ cmd/system-probe/subcommands/runtime/process_cache.go
@@ -1,0 +1,78 @@
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the Apache License Version 2.0.
+// This product includes software developed at Datadog (https://www.datadoghq.com/).
+// Copyright 2016-present Datadog, Inc.
+
+//go:build linux || windows
+
+// Package runtime holds runtime related files
+package runtime
+
+import (
+	"fmt"
+
+	"github.com/DataDog/datadog-agent/cmd/system-probe/command"
+	"github.com/DataDog/datadog-agent/comp/core"
+	"github.com/DataDog/datadog-agent/comp/core/config"
+	log "github.com/DataDog/datadog-agent/comp/core/log/def"
+	"github.com/DataDog/datadog-agent/comp/core/secrets"
+	secagent "github.com/DataDog/datadog-agent/pkg/security/agent"
+	"github.com/DataDog/datadog-agent/pkg/util/fxutil"
+	"github.com/spf13/cobra"
+	"go.uber.org/fx"
+)
+
+type processCacheDumpCliParams struct {
+	*command.GlobalParams
+
+	withArgs bool
+	format   string
+}
+
+func processCacheCommands(globalParams *command.GlobalParams) []*cobra.Command {
+	cliParams := &processCacheDumpCliParams{
+		GlobalParams: globalParams,
+	}
+
+	processCacheDumpCmd := &cobra.Command{
+		Use:   "dump",
+		Short: "dump the process cache",
+		RunE: func(_ *cobra.Command, _ []string) error {
+			return fxutil.OneShot(dumpProcessCache,
+				fx.Supply(cliParams),
+				fx.Supply(core.BundleParams{
+					ConfigParams: config.NewAgentParams("", config.WithConfigMissingOK(true)),
+					SecretParams: secrets.NewDisabledParams(),
+					LogParams:    log.ForOneShot("SYS-PROBE", "info", true)}),
+				core.Bundle(),
+			)
+		},
+	}
+	processCacheDumpCmd.Flags().BoolVar(&cliParams.withArgs, "with-args", false, "add process arguments to the dump")
+	processCacheDumpCmd.Flags().StringVar(&cliParams.format, "format", "dot", "process cache dump format")
+
+	processCacheCmd := &cobra.Command{
+		Use:   "process-cache",
+		Short: "process cache",
+	}
+	processCacheCmd.AddCommand(processCacheDumpCmd)
+
+	return []*cobra.Command{processCacheCmd}
+}
+
+func dumpProcessCache(_ log.Component, _ config.Component, _ secrets.Component, processCacheDumpArgs *processCacheDumpCliParams) error {
+	client, err := secagent.NewRuntimeSecurityClient()
+	if err != nil {
+		return fmt.Errorf("unable to create a runtime security client instance: %w", err)
+	}
+	defer client.Close()
+
+	filename, err := client.DumpProcessCache(processCacheDumpArgs.withArgs, processCacheDumpArgs.format)
+	if err != nil {
+		return fmt.Errorf("unable to get a process cache dump: %w", err)
+	}
+
+	fmt.Printf("Process dump file: %s\n", filename)
+
+	return nil
+}
