+++ cmd/system-probe/subcommands/runtime/selftest.go
@@ -1,0 +1,58 @@
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the Apache License Version 2.0.
+// This product includes software developed at Datadog (https://www.datadoghq.com/).
+// Copyright 2016-present Datadog, Inc.
+
+//go:build linux || windows
+
+package runtime
+
+import (
+	"fmt"
+
+	"github.com/DataDog/datadog-agent/cmd/system-probe/command"
+	"github.com/DataDog/datadog-agent/comp/core"
+	"github.com/DataDog/datadog-agent/comp/core/config"
+	log "github.com/DataDog/datadog-agent/comp/core/log/def"
+	"github.com/DataDog/datadog-agent/comp/core/secrets"
+	secagent "github.com/DataDog/datadog-agent/pkg/security/agent"
+	"github.com/DataDog/datadog-agent/pkg/util/fxutil"
+	"github.com/spf13/cobra"
+	"go.uber.org/fx"
+)
+
+func selfTestCommand(_ *command.GlobalParams) *cobra.Command {
+	return &cobra.Command{
+		Use:   "self-test",
+		Short: "Run runtime self test",
+		RunE: func(_ *cobra.Command, _ []string) error {
+			return fxutil.OneShot(runRuntimeSelfTest,
+				fx.Supply(core.BundleParams{
+					ConfigParams: config.NewAgentParams("", config.WithConfigMissingOK(true)),
+					SecretParams: secrets.NewDisabledParams(),
+					LogParams:    log.ForOneShot("SYS-PROBE", "info", true)}),
+				core.Bundle(),
+			)
+		},
+	}
+}
+
+func runRuntimeSelfTest(_ log.Component, _ config.Component, _ secrets.Component) error {
+	client, err := secagent.NewRuntimeSecurityClient()
+	if err != nil {
+		return fmt.Errorf("unable to create a runtime security client instance: %w", err)
+	}
+	defer client.Close()
+
+	selfTestResult, err := client.RunSelfTest()
+	if err != nil {
+		return fmt.Errorf("unable to get a process self test: %w", err)
+	}
+
+	if selfTestResult.Ok {
+		fmt.Printf("Runtime self test: OK\n")
+	} else {
+		fmt.Printf("Runtime self test: error: %v\n", selfTestResult.Error)
+	}
+	return nil
+}
