+++ go.mod
@@ -479,6 +479,7 @@ require (
 	github.com/mattn/go-sqlite3 v1.14.24
 	github.com/shirou/gopsutil/v4 v4.25.2
 	go.opentelemetry.io/collector/component/componenttest v0.121.0
+	gotest.tools v2.2.0+incompatible
 	modernc.org/sqlite v1.36.2
 )
